requirements.txt
discord.py==1.3.4
PlexAPI==4.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.0
